requirements.txt
Django==1.6.3
django-rest-swagger==0.1.14
djangorestframework==2.3.13
